refactor: extract mailbox finish wrappers and TurnFinishedSignal cluster (giant baselines 4168/3194)
